"Add support for environment variable ENABLE_ACKNOWLEDGMENT to enable optimized acknowledgment task in core alert pipeline. (#431)"
This commit adds a new feature that allows developers to enable or disable the use of an optimized acknowledgment task in the core alert pipeline by setting the `ENABLE_ACKNOWLEDGMENT` environment variable to `true` or `false`, respectively. The change also updates the code to use the correct syntax for creating an asynchronous task using `asyncio.create_task()` and adds logging statements to indicate when the optimized acknowledgment check is scheduled.

The commit message includes a reference to GitHub issue #431, which describes the problem that the new feature solves. It also includes a brief description of the changes made in the code, such as adding a new environment variable and updating the logic for scheduling the optimized acknowledgment task. Overall, the commit message provides enough information for developers to understand the purpose of the change and how it works.